Move column chunk stats to separate file
