+++ src/include/storage/store/column_chunk_data.h
@@ -12,6 +12,7 @@
 #include "storage/compression/compression.h"
 #include "storage/enums/residency_state.h"
 #include "storage/store/column_chunk_metadata.h"
+#include "storage/store/column_chunk_stats.h"
 #include "storage/store/column_reader_writer.h"
 #include "storage/store/in_memory_exception_chunk.h"
 
@@ -28,16 +29,6 @@ namespace storage {
 
 class Column;
 class NullChunkData;
-
-struct ColumnChunkStats {
-    std::optional<StorageValue> max;
-    std::optional<StorageValue> min;
-
-    void update(std::optional<StorageValue> min, std::optional<StorageValue> max,
-        common::PhysicalTypeID dataType);
-    void update(StorageValue val, common::PhysicalTypeID dataType);
-    void reset();
-};
 
 // TODO(bmwinger): Hide access to variables.
 struct ChunkState {
